+++ examples/complete/main.tf
@@ -1,0 +1,35 @@
+terraform {
+  required_providers {
+    netdata = {
+      # TODO: Update this string with the published name of your provider.
+      source = "netdata.cloud/todo/netdata"
+    }
+  }
+  required_version = ">= 1.1.0"
+}
+
+provider "netdata" {}
+
+resource "netdata_space" "test" {
+  name        = "MyTestingSpace"
+  description = "Created by Terraform"
+}
+
+resource "netdata_room" "test" {
+  spaceid     = netdata_space.test.id
+  name        = "MyTestingRoom"
+  description = "Created by Terraform2"
+}
+
+data "netdata_space" "test" {
+  id = netdata_space.test.id
+}
+
+data "netdata_room" "test" {
+  id      = netdata_room.test.id
+  spaceid = netdata_space.test.id
+}
+
+output "datasource" {
+  value = data.netdata_space.test.name
+}