+++ examples/data-sources/netdata_room/data-source.tf
@@ -1,0 +1,4 @@
+data "netdata_room" "test" {
+  spaceid = "<spaceid>"
+  id      = "<roomid>"
+}