+++ examples/resources/netdata_room/import.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+terraform import netdata_room.test spaceid,roomid