+++ examples/resources/netdata_room/resource.tf
@@ -1,0 +1,5 @@
+resource "netdata_room" "test" {
+  spaceid     = "<spaceid>"
+  name        = "MyTestingSpace"
+  description = "Created by Terraform"
+}