+++ internal/client/models.go
@@ -5,3 +5,9 @@ type SpaceInfo struct {
 	Name        string `json:"name"`
 	Description string `json:"description"`
 }
+
+type RoomInfo struct {
+	ID          string `json:"id"`
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}