+++ CHANGELOG.md
@@ -5,3 +5,4 @@ Initial version.
 FEATURES:
 
 - support for the Netdata Cloud Spaces
+- support for the Netdata Cloud Rooms